+++ improver/metadata/utilities.py
@@ -12,7 +12,7 @@ import dask.array as da
 import iris
 import numpy as np
 from cf_units import Unit
-from iris.cube import Cube
+from iris.cube import Cube, CubeList
 from numpy import ndarray
 from numpy.ma.core import MaskedArray
 
@@ -175,6 +175,43 @@ def create_coordinate_hash(cube: Cube) -> str:
             ]
         )
     return generate_hash(hashable_data)
+
+
+def check_grid_match(cubes: Union[List[Cube], CubeList]) -> None:
+    """
+    Checks that cubes are on, or originate from, compatible coordinate grids.
+    Each cube is first checked for an existing 'model_grid_hash' which can be
+    used to encode coordinate information on cubes that do not themselves
+    contain a coordinate grid (e.g. spotdata cubes). If this is not found a new
+    hash is generated to enable comparison. If the cubes are not compatible, an
+    exception is raised to prevent the use of unmatched cubes.
+
+    Args:
+        cubes:
+            A list of cubes to check for grid compatibility.
+
+    Raises:
+        ValueError: Raised if the cubes are not on matching grids as
+                    identified by the model_grid_hash.
+    """
+
+    def _get_grid_hash(cube):
+        try:
+            cube_hash = cube.attributes["model_grid_hash"]
+        except KeyError:
+            cube_hash = create_coordinate_hash(cube)
+        return cube_hash
+
+    cubes = iter(cubes)
+    reference_hash = _get_grid_hash(next(cubes))
+
+    for cube in cubes:
+        cube_hash = _get_grid_hash(cube)
+        if cube_hash != reference_hash:
+            raise ValueError(
+                "Cubes do not share or originate from the same "
+                "grid, so cannot be used together."
+            )
 
 
 def get_model_id_attr(cubes: List[Cube], model_id_attr: str) -> str:
