+++ improver/spotdata/apply_lapse_rate.py
@@ -14,7 +14,8 @@ from iris.exceptions import CoordinateNotFoundError
 from improver import PostProcessingPlugin
 from improver.lapse_rate import compute_lapse_rate_adjustment
 from improver.metadata.probabilistic import is_probability
-from improver.spotdata.spot_extraction import SpotExtraction, check_grid_match
+from improver.metadata.utilities import check_grid_match
+from improver.spotdata.spot_extraction import SpotExtraction
 
 
 class SpotLapseRateAdjust(PostProcessingPlugin):
